Add files via upload
Example file based upon the Revit demo project and Speckle transfer to Blender. Note that this example only works in IDA ICE 4.8 or earlier with relevant plugins. Also note that one of the zones are located outside of the building body and this should be adjusted in IDA ICE or in Blender before export. Also note that the code will probably change before final release.