Fixed typos and grammar in the Readme
Automatically and optionally was not spelled correctly. Corrected the grammar in a few cases. Still unsure what the sentence "- Windows and doors should have center of origin as the to center of mass (surface)" is suppose to mean. Is it "as the center of mass" or " as the top center of mass"?